# pkg/api/api.yml
openapi: 3.0.3
info:
  title: oidc-serv API
  version: 0.1.0
paths:
  /users:
    post:
      operationId: createUser
      summary: Create a user
      description: |
        Creates a new user account.

        - `username`: 3–32 characters, letters/digits/underscores only
        - `password`: minimum 8 characters; stored as a bcrypt hash, never returned
        - `locale`: BCP 47 language tag; defaults to `en` if omitted
      tags:
        - Users
      security:
        - sessionAuth: []
        - tmbAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/InternalError'
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserIdParam'
    get:
      operationId: getUserById
      summary: Get a user by ID
      tags:
        - Users
      security:
        - sessionAuth: []
        - tmbAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /users/by-username/{username}:
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
          minLength: 3
          maxLength: 32
    get:
      operationId: getUserByUsername
      summary: Get a user by username
      tags:
        - Users
      security:
        - sessionAuth: []
        - tmbAuth: []
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
  /me:
    get:
      operationId: getMe
      summary: Get the authenticated user
      tags:
        - Me
      security:
        - sessionAuth: []
        - tmbAuth: []
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateMe
      summary: Update the authenticated user's username and/or locale
      tags:
        - Me
      security:
        - sessionAuth: []
        - tmbAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '204':
          description: Updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteMe
      summary: Delete the authenticated user
      tags:
        - Me
      security:
        - sessionAuth: []
        - tmbAuth: []
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalError'
  /me/password:
    patch:
      operationId: updateMyPassword
      summary: Change the authenticated user's password
      description: |
        Verifies `old_password` against the stored bcrypt hash before applying the change.
        Returns `401` for both wrong `old_password` and missing authentication,
        using the same error to prevent information leakage.
      tags:
        - Me
      security:
        - sessionAuth: []
        - tmbAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '204':
          description: Password updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalError'
  /clients:
    post:
      operationId: createClient
      summary: Create an OIDC client
      description: |
        Creates a new OIDC client.

        - `id`: chosen by the caller; must be unique
        - `secret`: plaintext secret — generated if omitted; returned **only once**, store securely
        - `redirect_uris`: at least one URI is required
      tags:
        - Clients
      security:
        - adminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateClientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        default:
          $ref: '#/components/responses/InternalError'
  /clients/{clientId}:
    parameters:
      - $ref: '#/components/parameters/ClientIdParam'
    get:
      operationId: getClientById
      summary: Get an OIDC client by ID
      tags:
        - Clients
      security:
        - adminKeyAuth: []
      responses:
        '200':
          description: Client found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OIDCClient'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
    delete:
      operationId: deleteClient
      summary: Delete an OIDC client
      tags:
        - Clients
      security:
        - adminKeyAuth: []
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      description: |
        Session key issued by the login endpoint.
        Format: `Bearer <session-key>`
    tmbAuth:
      type: http
      scheme: bearer
      description: |
        Trust Me Bro — development/internal use only.
        Format: `Bearer TMB <user-uuid>`
    adminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: |
        Static admin API key configured via the `ADMIN_API_KEY` environment variable.
        Required for all client management endpoints.
  schemas:
    CreateUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: ^[a-zA-Z0-9_]+$
          example: alice
        password:
          description: Plain text; stored as bcrypt hash, never returned.
          type: string
          format: password
          minLength: 8
          example: s3cr3tpassword
        locale:
          description: BCP 47 language tag. Defaults to `en` if omitted.
          type: string
          example: en-US
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        code:
          type: string
        error:
          type: string
          example: resource not found
    CreateUserResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          example: 018e4f3a-1b2c-7d8e-9f0a-b1c2d3e4f5a6
    ResourceConflictError:
      type: object
      required:
        - code
        - error
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - properties:
            code:
              type: string
              enum:
                - resource_exists
              example: resource_exists
    User:
      description: Safe user projection. Never includes the password hash.
      type: object
      required:
        - id
        - username
        - locale
      properties:
        id:
          type: string
          format: uuid
          example: 018e4f3a-1b2c-7d8e-9f0a-b1c2d3e4f5a6
        username:
          type: string
          minLength: 3
          maxLength: 32
          example: alice
        locale:
          description: BCP 47 language tag.
          type: string
          example: en-US
    UpdateUserRequest:
      description: At least one field must be provided.
      type: object
      minProperties: 1
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: ^[a-zA-Z0-9_]+$
          example: alice
        locale:
          description: BCP 47 language tag.
          type: string
          example: en-US
    UpdatePasswordRequest:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          format: password
          minLength: 8
        new_password:
          type: string
          format: password
          minLength: 8
    CreateClientRequest:
      type: object
      required:
        - id
        - redirect_uris
      properties:
        id:
          type: string
          example: my-client
        secret:
          description: Plaintext secret. Generated if omitted.
          type: string
          format: password
        redirect_uris:
          type: array
          minItems: 1
          items:
            type: string
            format: uri
        grant_types:
          type: array
          items:
            type: string
        response_types:
          type: array
          items:
            type: string
        scope:
          type: string
          example: openid profile
        token_endpoint_auth_method:
          type: string
          default: client_secret_basic
    CreateClientResponse:
      type: object
      required:
        - id
        - secret
      properties:
        id:
          type: string
          example: my-client
        secret:
          description: Plaintext client secret — returned only once, store securely.
          type: string
    OIDCClient:
      description: OIDC client metadata. Never includes the secret.
      type: object
      required:
        - id
        - redirect_uris
        - grant_types
        - response_types
        - scope
        - token_endpoint_auth_method
        - created_at
        - expires_at
      properties:
        id:
          type: string
          example: my-client
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
        grant_types:
          type: array
          items:
            type: string
          example:
            - authorization_code
        response_types:
          type: array
          items:
            type: string
          example:
            - code
        scope:
          type: string
          example: openid profile
        token_endpoint_auth_method:
          type: string
          example: client_secret_basic
        created_at:
          type: integer
          format: int64
          description: Unix timestamp
        expires_at:
          type: integer
          format: int64
          description: Unix timestamp; 0 means no expiry
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceConflictError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
    UserIdParam:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        example: 018e4f3a-1b2c-7d8e-9f0a-b1c2d3e4f5a6
    ClientIdParam:
      name: clientId
      in: path
      required: true
      schema:
        type: string
        example: my-client